tetechJournal/telegram_agent/requirements.txt
# pipreqs . --force
python-dotenv==1.0.1
python-telegram-bot==21.10
crewai>=0.11.0
python-dotenv==1.0.1
utils==1.0.2
httpx
